test: converting a portion of tests from easymock to mockito
