+++ google-cloud-storage/src/test/java/com/google/cloud/storage/StorageImplTest.java
@@ -790,162 +790,6 @@ public class StorageImplTest {
   }
 
   @Test
-  public void testGetBucket() {
-    EasyMock.expect(storageRpcMock.get(BucketInfo.of(BUCKET_NAME1).toPb(), EMPTY_RPC_OPTIONS))
-        .andReturn(BUCKET_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Bucket bucket = storage.get(BUCKET_NAME1);
-    assertEquals(expectedBucket1, bucket);
-  }
-
-  @Test
-  public void testGetBucketWithOptions() {
-    EasyMock.expect(storageRpcMock.get(BucketInfo.of(BUCKET_NAME1).toPb(), BUCKET_GET_OPTIONS))
-        .andReturn(BUCKET_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Bucket bucket = storage.get(BUCKET_NAME1, BUCKET_GET_METAGENERATION);
-    assertEquals(expectedBucket1, bucket);
-  }
-
-  @Test
-  public void testGetBucketWithSelectedFields() {
-    Capture<Map<StorageRpc.Option, Object>> capturedOptions = Capture.newInstance();
-    EasyMock.expect(
-            storageRpcMock.get(
-                EasyMock.eq(BucketInfo.of(BUCKET_NAME1).toPb()), EasyMock.capture(capturedOptions)))
-        .andReturn(BUCKET_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Bucket bucket = storage.get(BUCKET_NAME1, BUCKET_GET_METAGENERATION, BUCKET_GET_FIELDS);
-    assertEquals(
-        BUCKET_GET_METAGENERATION.getValue(),
-        capturedOptions.getValue().get(BUCKET_GET_METAGENERATION.getRpcOption()));
-    String selector = (String) capturedOptions.getValue().get(BLOB_GET_FIELDS.getRpcOption());
-    assertTrue(selector.contains("name"));
-    assertTrue(selector.contains("location"));
-    assertTrue(selector.contains("acl"));
-    assertEquals(17, selector.length());
-    assertEquals(BUCKET_INFO1.getName(), bucket.getName());
-  }
-
-  @Test
-  public void testGetBucketWithEmptyFields() {
-    Capture<Map<StorageRpc.Option, Object>> capturedOptions = Capture.newInstance();
-    EasyMock.expect(
-            storageRpcMock.get(
-                EasyMock.eq(BucketInfo.of(BUCKET_NAME1).toPb()), EasyMock.capture(capturedOptions)))
-        .andReturn(BUCKET_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Bucket bucket = storage.get(BUCKET_NAME1, BUCKET_GET_METAGENERATION, BUCKET_GET_EMPTY_FIELDS);
-    assertEquals(
-        BUCKET_GET_METAGENERATION.getValue(),
-        capturedOptions.getValue().get(BUCKET_GET_METAGENERATION.getRpcOption()));
-    String selector = (String) capturedOptions.getValue().get(BLOB_GET_FIELDS.getRpcOption());
-    assertTrue(selector.contains("name"));
-    assertEquals(4, selector.length());
-    assertEquals(BUCKET_INFO1.getName(), bucket.getName());
-  }
-
-  @Test
-  public void testGetBlob() {
-    EasyMock.expect(
-            storageRpcMock.get(BlobId.of(BUCKET_NAME1, BLOB_NAME1).toPb(), EMPTY_RPC_OPTIONS))
-        .andReturn(BLOB_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Blob blob = storage.get(BUCKET_NAME1, BLOB_NAME1);
-    assertEquals(expectedBlob1, blob);
-  }
-
-  @Test
-  public void testGetBlobWithOptions() {
-    EasyMock.expect(
-            storageRpcMock.get(BlobId.of(BUCKET_NAME1, BLOB_NAME1).toPb(), BLOB_GET_OPTIONS))
-        .andReturn(BLOB_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Blob blob = storage.get(BUCKET_NAME1, BLOB_NAME1, BLOB_GET_METAGENERATION, BLOB_GET_GENERATION);
-    assertEquals(expectedBlob1, blob);
-  }
-
-  @Test
-  public void testGetBlobWithOptionsFromBlobId() {
-    EasyMock.expect(storageRpcMock.get(BLOB_INFO1.getBlobId().toPb(), BLOB_GET_OPTIONS))
-        .andReturn(BLOB_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Blob blob =
-        storage.get(
-            BLOB_INFO1.getBlobId(), BLOB_GET_METAGENERATION, BLOB_GET_GENERATION_FROM_BLOB_ID);
-    assertEquals(expectedBlob1, blob);
-  }
-
-  @Test
-  public void testGetBlobWithSelectedFields() {
-    Capture<Map<StorageRpc.Option, Object>> capturedOptions = Capture.newInstance();
-    EasyMock.expect(
-            storageRpcMock.get(
-                EasyMock.eq(BlobId.of(BUCKET_NAME1, BLOB_NAME1).toPb()),
-                EasyMock.capture(capturedOptions)))
-        .andReturn(BLOB_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Blob blob =
-        storage.get(
-            BUCKET_NAME1,
-            BLOB_NAME1,
-            BLOB_GET_METAGENERATION,
-            BLOB_GET_GENERATION,
-            BLOB_GET_FIELDS);
-    assertEquals(
-        BLOB_GET_METAGENERATION.getValue(),
-        capturedOptions.getValue().get(BLOB_GET_METAGENERATION.getRpcOption()));
-    assertEquals(
-        BLOB_GET_GENERATION.getValue(),
-        capturedOptions.getValue().get(BLOB_GET_GENERATION.getRpcOption()));
-    String selector = (String) capturedOptions.getValue().get(BLOB_GET_FIELDS.getRpcOption());
-    assertTrue(selector.contains("bucket"));
-    assertTrue(selector.contains("name"));
-    assertTrue(selector.contains("contentType"));
-    assertTrue(selector.contains("crc32c"));
-    assertEquals(30, selector.length());
-    assertEquals(expectedBlob1, blob);
-  }
-
-  @Test
-  public void testGetBlobWithEmptyFields() {
-    Capture<Map<StorageRpc.Option, Object>> capturedOptions = Capture.newInstance();
-    EasyMock.expect(
-            storageRpcMock.get(
-                EasyMock.eq(BlobId.of(BUCKET_NAME1, BLOB_NAME1).toPb()),
-                EasyMock.capture(capturedOptions)))
-        .andReturn(BLOB_INFO1.toPb());
-    EasyMock.replay(storageRpcMock);
-    initializeService();
-    Blob blob =
-        storage.get(
-            BUCKET_NAME1,
-            BLOB_NAME1,
-            BLOB_GET_METAGENERATION,
-            BLOB_GET_GENERATION,
-            BLOB_GET_EMPTY_FIELDS);
-    assertEquals(
-        BLOB_GET_METAGENERATION.getValue(),
-        capturedOptions.getValue().get(BLOB_GET_METAGENERATION.getRpcOption()));
-    assertEquals(
-        BLOB_GET_GENERATION.getValue(),
-        capturedOptions.getValue().get(BLOB_GET_GENERATION.getRpcOption()));
-    String selector = (String) capturedOptions.getValue().get(BLOB_GET_FIELDS.getRpcOption());
-    assertTrue(selector.contains("bucket"));
-    assertTrue(selector.contains("name"));
-    assertEquals(11, selector.length());
-    assertEquals(expectedBlob1, blob);
-  }
-
-  @Test
   public void testListBuckets() {
     String cursor = "cursor";
     ImmutableList<BucketInfo> bucketInfoList = ImmutableList.of(BUCKET_INFO1, BUCKET_INFO2);
